md5Digest(): explicitly provide getBytes() encoding
If encoding not provided, String.getBytes() uses system encoding to convert string into array of bytes. If system encoding is not set or it is not unicode, then unicode symbols may be lost/altered (we experienced this in our docker container). As a result, md5 checksum will be wrong (I'm not sure if the message itself gets corrupted). This should fix #13.