151: Fixing long polling FIFO requests using a requestAttemptId
The fifo request attempt cache is being populated by an empty list which
would cause subsequent attempts to retrieve messages to always return an
empty list as it would never go back to the queue. This change ensures
that if the cache returns an empty list an attempt to dequeue a message
will utilize the message queue itself rather than the cache.